Split the conditions from where
